workers/requirements.txt
numpy>=1.26
pandas>=2.2
scipy>=1.13
statsmodels>=0.14
yfinance>=0.2.40
psycopg[binary]>=3.2
sqlalchemy>=2.0
python-dotenv>=1.0
APScheduler>=3.10
requests>=2.32
lxml>=5.2
